% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_comp_from_mz.R
\name{assign_comp_from_mz}
\alias{assign_comp_from_mz}
\title{Assign elemental composition to m/z value}
\usage{
assign_comp_from_mz(
  mz,
  elements,
  isotope,
  mins.list,
  maxs.list,
  ionmode,
  error.ppm = 5,
  rdbrange = c(-1, 100),
  rdb.rule = "none"
)
}
\arguments{
\item{mz}{list of m/z values to which composition is to be assigned}

\item{elements}{list of elements which may appear in composition (Exact element letter only, case sensitive). E.g c("H", "C", "Cl")}

\item{isotope}{atomic number or isotope number of desired element (NOT atomic mass). E.g.  c(1, 12, 35)}

\item{mins.list}{minimum number of elements which are allowed in composition in the same order as elements and isotope arguments.}

\item{maxs.list}{maximum number of elements which are allowed in composition in the same order as elements and isotope arguments.}

\item{ionmode}{Ion mode the samples were run in. Must be one of "neg.ion", "pos.ion", or "neutral"}

\item{error.ppm}{the maximum error allowed for assigned compositions. Defaults to 5.}

\item{rdbrange}{the range of allowed RDB (Relative double bond equivalents) values of assigned compositons. Defaults to c(-1,100)}

\item{rdb.rule}{rule to force assignment of elemental compositions to only those that have a certain RDB characteristic (e.g integer values or non-integer values). Options are "none", "nonint", or "int". Defaults to "none" meaning composition will be assigned as the closest match to m/z value regardless of its RDB.}
}
\description{
This function assigns elemental composition to user-supplied m/z values in accordance with user-supplied constraints
}
\examples{
df$gen.structures <- assign_structures(comp = df$composition,
                                       ion.mode = "neg.ion",
                                       domain = "euk",
                                       max.dbl.bnds = 8)
}
\keyword{assignment}
\keyword{composition}
\keyword{elemental}
