% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_structures.R
\name{assign_structures}
\alias{assign_structures}
\title{Assign structures to a list of elemental compositions}
\usage{
assign_structures(
  comp,
  ion.mode,
  adducts = NULL,
  domain = "euk",
  lois = NULL,
  max.dbl.bnds = 14,
  carbon_isotope_symbol = "C[13]",
  hydrogen_isotope_symbol = "D[2]"
)
}
\arguments{
\item{comp}{list (or dataframe column) of elemental compositions for which structures are to be assigned.}

\item{ion.mode}{which ion mode the samples were run in. One of "neg.ion","pos.ion", or "neutral" ("neutral" searches for structures as exact compositions / neutral species).}

\item{adducts}{which ions/adducts to search for compositions as. Defaults to c("m.plus.h","m.plus.ammonia","m.plus.sodium") in positive ion mode, c("m.minus.h","m.plus.chloride","m.minus.2h") in negative ion mode, and the inclusive union of these two lists in neutral mode.}

\item{domain}{what domain of life the data originates from. "euk" for eukaryotic (default) or "bact" for bacterial.}

\item{lois}{which lipid classes to look for. Defaults to c("pc","sm","tag","dag","gdg") in positive ion mode, c("pe","cer","cl","pi","pg","pa","ps","ffa","gdg") in negative ion mode, and the inclusive union of these two lists in neutral mode (Diacylglycerol = "dag", Triacylglycerol = "tag", Phosphatidic acid = "pa", Phosphatidylglycerol = "pg", Phosphatidylinositol = "pi", Cardiolipin = "cl", Sphingomyelin = "sm", Ceramide = "cer", Phosphatidylcholine = "pc", Phosphatidylethanolamine = "pe", Phosphatidylserine = "ps", Free fatty acids = "ffa", Glycosyldiacylglycerols = "gdg")}

\item{max.dbl.bnds}{maximum number of acyl+alkyl chain double bonds allowed in returned structures. Defaults to 14.}

\item{carbon_isotope_symbol}{string which is used to denote carbon isotope used, if applicable. Defaults to "C[13]".}

\item{hydrogen_isotope_symbol}{string which is used to denote hydrogen isotope used, if applicable. Defaults to "D[2]".}
}
\description{
This function assigns general lipid structure (e.g. PC(32:1)) to a user-defined elemental composition (e.g. C40 H79 N O8 P), or a list/column of elemental compositions
}
\examples{
df$gen.structures <- assign_structures(comp = df$composition,
                                       ion.mode = "neg.ion",
                                       domain = "euk",
                                       max.dbl.bnds = 8)
}
\keyword{assignment}
\keyword{structural}
