% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_num_elements.R
\name{extract_num_elements}
\alias{extract_num_elements}
\title{Extract number of a given element}
\usage{
extract_num_elements(element_letter, column_of_df)
}
\arguments{
\item{element_letter}{The element letter you would like to return the number of. Case does not matter. Allowed options are c("C","H","O","N","P","Na","Cl")}

\item{column_of_df}{A list (ideally, a column of a dataframe) which you'd like to extract the number of elements from}
}
\description{
This function extracts the number of a single element from each composition in a column of compositions
}
\examples{
cat_function("C",df$composition)
}
\keyword{composition}
