% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_exact_structures.R
\name{guess_exact_structures}
\alias{guess_exact_structures}
\title{Guess exact structures from general structures}
\usage{
guess_exact_structures(structures)
}
\arguments{
\item{structures}{list (or columns) containing the general structures (from assign_structures() function) for which exact structures are to be queried}
}
\description{
This function gueses an exact lipid structure  (e.g. 16:0/18:1-PE) from general lipid structures (e.g. PE(34:1)) which come from the assign_structures() function
}
\examples{
df$structures <- assign_structures(df$c,df$h,df$o,df$n,df$p,df$na,df$cl,
                  "neg.ion",
                  "m.minus.h",
                  "euk",
                   c("pe","pa","pg","ffa"),
                   6)
                   
exact_structures_df <- guess_exact_structures(df$structures)

df <- cbind(df,exact_structures_df)
}
\keyword{assignment}
\keyword{structural}
