% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_sl.R
\name{normalize_sl}
\alias{normalize_sl}
\title{Normalize shotgun lipidomics data matrix}
\usage{
normalize_sl(m, log.normalize = T)
}
\arguments{
\item{m}{a matrix of shotgun lipidomics data where rownames are lipid compositions, structures, or any other unique identifier, columns correspond to samples or replicates, and all entries of matrix are coercible to numeric}

\item{log.normalize}{log2 transform data before normalization? Default = TRUE}
}
\description{
This function performs a quantile normalization (preprocessCore::normalize.quantiles) of a shotgun lipidomics data matrix of 100 rows (= species) or more
}
\examples{
m <- df \%>\% dplyr::select(c("rel.intensity","sample","comp")) \%>\% 
            tidyr::pivot_wider(names_from = sample,
                               values_from = rel.intensity,
                               values_fn = mean))

m.norm <- normalize_sl(m)
}
