% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_comp_from_mz.R
\name{assign_comp_from_mz}
\alias{assign_comp_from_mz}
\title{Assign elemental composition to m/z value}
\usage{
assign_comp_from_mz(
  mz,
  elements,
  isotope,
  mins.list,
  maxs.list,
  ionmode,
  error.ppm = 5,
  rdbrange = c(-1, 100),
  rdb.rule = "none"
)
}
\arguments{
\item{mz}{list (or dataframe column) of m/z values to which composition is to be assigned}

\item{elements}{list of elements which may be included in composition (Exact element letter only, case sensitive).}

\item{isotope}{atomic number/isotope number of desired element (NOT atomic mass).}

\item{mins.list}{minimum number of elements which are allowed in composition (must be in the same order as elements and isotope arguments).}

\item{maxs.list}{maximum number of elements which are allowed in composition (must be in the same order as elements and isotope arguments).}

\item{ionmode}{ion mode the samples were run in. Must be one of "neg.ion", "pos.ion", or "neutral".}

\item{error.ppm}{the maximum error ("delta", in units ppm) allowed for assigned compositions. Defaults to 5 ppm.}

\item{rdbrange}{the range of allowed RDB (relative double bond equivalents) values of assigned compositions. Defaults to -1 to 100 (denoted: c(-1,100)).}

\item{rdb.rule}{force assignment of elemental compositions to only those that have a certain RDB characteristic (e.g integer values or non-integer values)? Options are "none", "nonint", or "int". Defaults to "none" meaning composition will be assigned as the closest match to m/z value regardless of RDB.}
}
\description{
This function assigns elemental composition to user-supplied m/z values in accordance with user-supplied constraints
}
\examples{
assign_comp_from_mz(df$mz,
                    elements = c("C","H","O","N","P","Na","Cl"),
                    isotope = c(12,1,16,14,31,23,35),
                    mins.list = c(10,20,2,0,0,0,0),
                    maxs.list = c(60,160,16,2,2,0,1),
                    ion.mode = "neg.ion",
                    error.ppm = 1,
                    rdb.rule = "nonint")

}
\keyword{assignment}
\keyword{composition}
\keyword{elemental}
