% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_structures.R
\name{assign_structures}
\alias{assign_structures}
\title{Assign structures to a list of compositions}
\usage{
assign_structures(
  c,
  h,
  o,
  n,
  p,
  na = NULL,
  cl = NULL,
  ion.mode,
  adducts = NULL,
  domain = "euk",
  lois = NULL,
  max.dbl.bnds = 14
)
}
\arguments{
\item{c, h, o, n, p}{lists (or columns) containing the numbers of each respective elemennt in each respective composition (these should come from from extract_num_elements() function)}

\item{na, cl}{lists (or columns) containing the numbers of each respective elemennt in each respective composition. Default to zero.}

\item{ion.mode}{which ion mode the samples were run in. One of "neg.ion","pos.ion",or "neutral" (neutral searches for structures as exact compositions / neutral species)}

\item{adducts}{which adducts to search for compositions as. Defaults to c("m.plus.h","m.plus.ammonia","m.plus.sodium") in positive ion mode, c("m.minus.h","m.plus.chloride","m.minus.2h") in negative ion mode, and the inclusive union of these two lists in neutral mode}

\item{domain}{what domain of life the data originates from. "euk" for eukaryotic, "bact" for bacterial.}

\item{lois}{which lipids to look for. Defaults to c("pc","sm","tag","dag") in positive ion mode, c("pe","cer","cl","pi","pg","pa","ps","ffa") in negative ion mode, and the inclusive union of these two lists in neutral mode.}

\item{max.dbl.bnds}{maximum acyl+alkyl chain double bonds allowed in returned structures. Defaults to 14.}
}
\description{
This function assigns general lipid structure (e.g. PC(32:1)) to a user-defined elemental composition
}
\examples{
assign_structures(df$c,df$h,df$o,df$n,df$p,df$na,df$cl,
                  "neg.ion",
                  "m.minus.h",
                  "euk",
                   c("pe","pa","pg","ffa"),
                   6)
}
\keyword{assignment}
\keyword{structureal}
