% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_num_elements.R
\name{extract_num_elements}
\alias{extract_num_elements}
\title{Extract the number of a given element from elemental compositions}
\usage{
extract_num_elements(element_letter, column_of_df)
}
\arguments{
\item{element_letter}{The element letter you would like to extract the number of. Case does not matter. Supported options are c("C","H","O","N","P","Na","Cl","[13C]","[2H]"), but the regex should be pretty flexible--just double check your results if using something other than these 9 options explicitly.}

\item{column_of_df}{A list (e.g. a column of a dataframe) of compositions which you'd like to extract the number of elements from.}
}
\description{
This function extracts the number of a single element (e.g. carbon, "C") from each composition (e.g. C43 H67 O8 N P) in a column of compositions
}
\examples{
df$c <- extract_num_elements("C",df$composition)
# for isotopes, make sure to pass the exact format of your isotope symbol to the element_letter argument
df$c_isotope <-  extract_num_elements("[13C]",df$composition)
df$h <- extract_num_elements("H",df$composition)
df$h_isotope <- extract_num_elements("[2H]",df$composition)
df$o <- extract_num_elements("O",df$composition)
df$n <- extract_num_elements("N",df$composition)
df$p <- extract_num_elements("P",df$composition)
df$na <- extract_num_elements("Na",df$composition)
df$cl <- extract_num_elements("Cl",df$composition)
}
\keyword{composition}
