% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_for_da_sl.R
\name{test_for_da_sl}
\alias{test_for_da_sl}
\title{Test all lipid species for differential abundance}
\usage{
test_for_da_sl(m, a, b)
}
\arguments{
\item{m}{a matrix of shotgun lipidomics data where rownames are lipid compositions, structures, or any other unique identifier, columns correspond to samples or replicates, and all entries of matrix are coercible to numeric (matrix should already be normalized and have had missing values imputted)}

\item{a}{column names corresponding to control condition}

\item{b}{column names corresponding to experimental condition}
}
\description{
This function performs tests all species for differential abundance across two biological conditions
}
\examples{
m <- df \%>\% dplyr::select(c("rel.intensity","sample","comp")) \%>\% 
            tidyr::pivot_wider(names_from = sample,
                               values_from = rel.intensity,
                               values_fn = mean))

m.norm <- normalize_sl(m)

m.final <- impute_sl(m.norm)

results <- test_for_da_sl(m.final, 
                          c("wt1","wt2","wt3"), 
                          c("ko1","ko2","ko3"))
}
\keyword{testing}
